Add --no-clean to checker (keeps the generated files around)

+++ tools/checker.py
@@ -691,7 +691,9 @@ def ParseArguments():
   parser.add_argument("--list-groups", dest="list_groups", action="store_true",
                       help="print a list of all groups found in the test output")
   parser.add_argument("--dump-group", dest="dump_group", metavar="GROUP",
-                      help="print the contents of an output group")
+                      help="print the contents of an output group"),
+  parser.add_argument("--no-clean", dest="no_clean", action="store_true",
+                      help="don't clean up generated files")
   return parser.parse_args()
 
 
@@ -773,4 +775,7 @@ if __name__ == "__main__":
     else:
       RunChecks(args.check_prefix, args.test_file, outputFile)
   finally:
-    shutil.rmtree(tempFolder)
+    if args.no_clean:
+      print("Files left in %s" % tempFolder)
+    else:
+      shutil.rmtree(tempFolder)
